requirements/requirements.txt
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2
pydantic==2.10.6
ruamel.yaml==0.18.6
scikit-learn==1.5.1
strictyaml==1.6.2